requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
sqlalchemy
requests
jinja2
markdown